wlserver: Move commit queues to per-xwayland server, rather than global
