+++ src/main.cpp
@@ -454,18 +454,3 @@ static bool initOutput( int preferredWidth, int preferredHeight, int preferredRe
 		return init_drm( &g_DRM, preferredWidth, preferredHeight, preferredRefresh );
 	}
 }
-
-void wayland_commit(struct wlr_surface *surf, struct wlr_buffer *buf)
-{
-	{
-		std::lock_guard<std::mutex> lock( wayland_commit_lock );
-
-		ResListEntry_t newEntry = {
-			.surf = surf,
-			.buf = buf,
-		};
-		wayland_commit_queue.push_back( newEntry );
-	}
-
-	nudge_steamcompmgr();
-}
